extract create*Hint method, use hint string in aggregator
