+++ packages/ovm/src/index.ts
@@ -1,3 +1,5 @@
+import * as hint from './hintString'
+
 export * from './types'
 export * from './DeciderManager'
 export * from './deciders'
@@ -5,3 +7,4 @@ export * from './decompiler'
 export * from './load'
 export * from './helpers'
 export * from './propertyFilter'
+export { hint }
